Enforce whitespaciness/emptiness of WhitespaceOrEmptyString
Update regex in WhitespaceOrEmptyString constructor to
only match strings with only whitespace or nothing in them,
and add tests.

It seems from the specification (rather than the XML schema) that
the intention of this type is only to allow for optional
ResourceReferences.

As it is not used in the module, also stop exporting the type from
QuakeML.Types.